requirements.txt
webexteamssdk
jinja2
requests
hurry.filesize